Resolve bug when thread has no member IDs
If a thread had no members in it (i.e. the user kicked everyone and then
removed themselves), there would be no line in the .htm with a comma
delimited list of the current members. The beautiful soup call finding
that line would return a none object and it would crash when trying to
run .split() on a none object. Also moved the regex and @facebook.com
string outside the class.